requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
llama-cpp-python==0.2.38
httpx==0.26.0
websockets==12.0
structlog==24.1.0
pytest==7.4.4
pytest-asyncio==0.23.3
rich==13.7.0
typer==0.9.0
aiohttp==3.9.3
tqdm==4.66.2
